Split Transactional in two
